pyproject.toml
[tool.poetry]
name = "eratos-xarray"
version = "0.2.0"
description = "Xarray backend for Eratos SDK"
authors = ["Chris Sharman <chris.sharman@csiro.au>"]
readme = "README.md"
packages = [{ include = "eratos_xarray" }]
license = "MIT"
homepage = "https://bitbucket.csiro.au/projects/SC/repos/eratos-xarray/browse"
repository = "https://bitbucket.csiro.au/projects/SC/repos/eratos-xarray/browse"

[tool.poetry.dependencies]
python = "^3.8"
xarray = "^2023.1.0"
numpy = "^1.20.0"
eratos-sdk = ">=0.18.0,<1.0.0"

[tool.poetry.plugins."xarray.backends"]
eratos = "eratos_xarray.backend.eratos_:EratosBackendEntrypoint"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
